# swagger.yaml
openapi: 3.0.3
info:
  title: Swagger API Course FE
  description: |-
    This is a list api of course FE
  contact:
    email: lttd0204@gmail.com
  version: 1.0.11
servers:
  - url: https://api-shop-l5ir.onrender.com
...
paths:
...
  /api/payment-type/{paymentTypeId}:
    get:
      tags:
        - payment-type
      summary: Get details payment type by ID
      description: Get details payment type by id of payment types
      operationId: getPaymentTypeById
      parameters:
        - name: paymentTypeId
          in: path
          description: ID of payment type that needs to be fetched
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/PaymentType"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/PaymentType"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
    delete:
      tags:
        - payment-type
      summary: Delete payment type by ID
      description: Try delete payment type by id of payment
      operationId: deletePaymentType
      parameters:
        - name: paymentTypeId
          in: path
          description: ID of the payment type that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/PaymentType"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/PaymentType"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
    put:
      tags:
        - payment-type
      summary: Update an existing payment type
      description: Update an existing payment type by Id
      operationId: updateCity
      parameters:
        - name: paymentTypeId
          in: path
          description: payment type id of payment type updated
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent payment type in the store
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentType"
          application/xml:
            schema:
              $ref: "#/components/schemas/PaymentType"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/PaymentType"
        required: true

      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/PaymentType"
            application/xml:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/PaymentType"
        "401":
          description: UNAUTHORIZED
        "500":
          description: INTERNAL_SERVER_ERROR
        "409":
          description: ALREADY_EXIST
        "400":
          description: INVALID
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    PaymentType:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
      required:
        - name
        - type
      xml:
        name: PaymentType
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT